second attempt; check_param flag.  Compiles, works as well as before... but doesn't seem to solve the mysteries I expected it to.

+++ dynchk/libs/libc/shutdown.c
@@ -6,19 +6,21 @@
 #undef shutdown
 static int(*funcptr) (int , int ) = 0;
 
+extern int __lsb_check_params;
 int shutdown (int arg0 , int arg1 )
 {
+	int reset_flag = __lsb_check_params;
+	int ret_value  ;
 	if(!funcptr)
 		funcptr = dlsym(RTLD_NEXT, "shutdown");
-	validate_NULL_TYPETYPE(  arg0, "shutdown - arg0");
-	validate_NULL_TYPETYPE(  arg1, "shutdown - arg1");
-	return funcptr(arg0, arg1);
-}
-
-int __lsb_shutdown (int arg0 , int arg1 )
-{
-	if(!funcptr)
-		funcptr = dlsym(RTLD_NEXT, "shutdown");
-	return funcptr(arg0, arg1);
+	if(__lsb_check_params)
+	{
+		__lsb_check_params=0;
+		validate_NULL_TYPETYPE(  arg0, "shutdown - arg0");
+		validate_NULL_TYPETYPE(  arg1, "shutdown - arg1");
+	}
+	ret_value = funcptr(arg0, arg1);
+	__lsb_check_params = reset_flag;
+	return ret_value;
 }
